translator_agents/pyproject.toml
[project]
name = "translator-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.18",
    "python-dotenv>=1.1.0",
]
